[core] payload, classifier, handler refactor
Rename entity to resource: This provides added clarity, since entity is vague.  Amazon references a ‘service’ as an offering like Kinesis or S3, while a Resource is an instance of the service, such as a Kinesis Stream, or a S3 bucket.

Move service/resource loading into Payload: This reduces the amount of work in the classifier class, and moves all payload attributes into the payload class.  The classifier uses attributes on the payload class to understand the map of all possible log types (and their attributes) to classify against.

Remove unnecessary debug logger statements: We capture these messages in metrics, so they aren’t needed anymore.